requirements.txt
aiofiles==24.1.0
aiogram==3.13.1
aiohappyeyeballs==2.4.2
aiohttp==3.10.6
aiosignal==1.3.1
aiosqlite==0.20.0
annotated-types==0.7.0
attrs==24.2.0
certifi==2024.8.30
frozenlist==1.4.1
idna==3.10
magic-filter==1.0.12
multidict==6.1.0
pip==24.2
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.0.1
SQLAlchemy==2.0.35
typing_extensions==4.12.2
yarl==1.13.0